# docs/architecture/openapi.yaml
openapi: 3.0.3
info:
  title: Cornven MVP API
  version: 0.3.0
  description: Contract-first boundary for POS raw records, deterministic Rental-rate settlement,
    reconciliation, and the internal read-only assistant. Customer repository fields remain isolated
    behind the adapter.
servers:
  - url: http://127.0.0.1:3119
    description: Isolated local integration API
paths:
...
  /api/v1/assistant:
    post:
      operationId: askAssistant
      summary: Classify one request with GPT-5.4 nano and execute a read-only tool
      description:
        Optional bounded conversation history supports clarification. No client permissions or
        arbitrary tool calls. Requires server OPENAI_API_KEY. Saved report reads do not recalculate.
        SOP answers require validated source citations.
      requestBody: &a12
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssistantRequest'
      responses:
        '200':
          description: Answer, clarification, refusal or tool error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AssistantAnswer'
        '400': &a1
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401': &a2
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '403': &a3
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404': &a4
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409': &a5
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '410': &a6
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500': &a7
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '502':
          description: Invalid model response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503': &a8
          description: ErrorEnvelope; report errors follow monthly-report-integration.zh-CN.md
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: Model timeout
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /api/v1/documents/status:
    get:
      operationId: getDocumentModuleStatus
      summary: Report whether approved SOP documents are configured
      description:
        Requires document permission tags injected by the trusted authentication or Assistant
        boundary. Requests without an authorised server context are denied.
      responses:
        '200':
          description: Module status
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                  - indexedDocuments
                  - indexedChunks
                  - embeddingProvider
                  - message
                properties:
                  status:
                    type: string
                    enum:
                      - not_configured
                      - ready
                  indexedDocuments:
                    type: integer
                    minimum: 0
                  indexedChunks:
                    type: integer
                    minimum: 0
                  embeddingProvider:
                    type: string
                  message:
                    type: string
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/documents/search:
    post:
      operationId: searchDocuments
      summary: Search authorised SOP chunks for citation-ready evidence
      description:
        The request cannot choose its own permissions. The server injects the caller's approved
        permission tags before retrieval.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentSearchInput'
      responses:
        '200':
          description: Permission-filtered document evidence
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSearchOutput'
        '400':
          $ref: '#/components/responses/ValidationError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '502':
          description: RAG_EVIDENCE_INVALID; source or citation integrity failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '503':
          description: RAG_CONFIG_INVALID, RAG_INDEX_UNAVAILABLE, RAG_PROVIDER_UNAVAILABLE or
            RAG_CAPACITY_EXCEEDED
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '504':
          description: RAG_PROVIDER_TIMEOUT; no automatic retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/assistant/tools/settlement.get:
    post:
      operationId: readSavedSettlement
      summary: Read the latest saved monthly report without calculating
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SavedSettlementQuery'
      responses:
        '200':
          description: Saved report or null
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavedSettlementResponse'
        '400': *a1
        '401': *a2
        '403': *a3
        '404': *a4
        '409': *a5
        '410': *a6
        '500': *a7
        '503': *a8
...
  /api/v1/assistant/tools/sales.refunds:
    post:
      operationId: readMonthlyRefundMetrics
      summary: Read monthly refund item units, transactions and amount without settlement calculation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MonthlyRefundMetricsRequest'
      responses:
        '200':
          description:
            Refund event metrics; quantity is positive item units, transaction count is distinct
            sourceTransactionId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonthlyRefundMetricsResponse'
        '400': *a1
        '401': *a2
        '403': *a3
        '404': *a4
        '409': *a5
        '410': *a6
        '500': *a7
        '503': *a8
  /api/v1/approvals:
    get:
      operationId: listMonthlyReportApprovals
      summary: List saved report versions by review status
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: &a34
              - draft
              - pending
              - approved
              - rejected
        - name: artistQuery
          in: query
          required: false
          schema:
            type: string
            maxLength: 200
        - name: settlementMonth
          in: query
          required: false
          schema:
            type: string
            pattern: ^\d{4}-(0[1-9]|1[0-2])$
        - name: cursor
          in: query
          required: false
          schema:
            type: string
            maxLength: 2000
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Paginated review queue
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalListResponse'
        '400': *a1
        '401': *a2
        '403': *a3
        '404': *a4
        '409': *a5
        '410': *a6
        '500': *a7
        '503': *a8
  /api/v1/approvals/{reportId}:
    get:
      operationId: getMonthlyReportApproval
      summary: Read saved snapshot, review status and audit history
      parameters:
        - *a9
      responses:
        '200':
          description: Version-specific approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDetail'
        '400': *a1
        '401': *a2
        '403': *a3
        '404': *a4
        '409': *a5
        '410': *a6
        '500': *a7
        '503': *a8
  /api/v1/approvals/{reportId}/actions:
    post:
      operationId: actOnMonthlyReportApproval
      summary: Submit, approve or reject a saved report version
      description: Rejection requires a nonblank reason. requestId is an idempotency key per report
        version; reuse with different content returns 409. Only draft -> pending ->
        approved/rejected is allowed. Actor comes from server identity. Does not modify report
        JSON/PDF or execute payment.
      parameters:
        - *a9
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApprovalCommand'
      responses:
        '200':
          description: Current status and history after action or replay
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApprovalDetail'
        '400': *a1
        '401': *a2
        '403': *a3
        '404': *a4
        '409': *a5
        '410': *a6
        '500': *a7
        '503': *a8
...
  /api/v1/assistant/stream:
    post:
      operationId: streamAssistant
      summary: Stream assistant progress and provisional SOP answer text over SSE
      description:
        POST with the same input as askAssistant. SSE data is an AssistantStreamEvent JSON
        object; event names are status, text_delta, final and error. Only final contains a validated
        answer with citations/results and replaces provisional text. A stream without final/error is
        incomplete. Disconnect cancels model work; no automatic retry or replay. Business answers
        are deterministic and delivered in final. Pre-stream failures use HTTP ErrorEnvelope; later
        failures use an error event. Comment heartbeats may occur.
      requestBody: *a12
      responses:
        '200':
          description: SSE events with JSON payloads described by AssistantStreamEvent
          content:
            text/event-stream:
              schema:
                type: string
        '400': *a1
        '401': *a2
        '403': *a3
        '404': *a4
        '409': *a5
        '410': *a6
        '500': *a7
        '503': *a8
...
components:
  responses:
    ValidationError:
      description: Request failed schema or business validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ForbiddenError:
      description: The authenticated actor is not allowed to access the protected resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    AssistantRequest:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
        history:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - role
              - content
            properties:
              role:
                type: string
                enum:
                  - user
                  - assistant
              content:
                type: string
                minLength: 1
                maxLength: 4000
...
    DocumentSearchInput:
      type: object
      additionalProperties: false
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          maxLength: 4000
          pattern: \S
        limit:
          type: integer
          minimum: 1
          maximum: 15
          default: 5
    DocumentSearchResult:
      type: object
      additionalProperties: false
      required:
        - documentId
        - documentVersion
        - chunkId
        - title
        - locator
        - excerpt
        - content
      properties:
        documentId:
          type: string
        documentVersion:
          type: string
        chunkId:
          type: string
        title:
          type: string
        locator:
          type: string
          description: Stable document, chapter, subsection, and optional part location.
        excerpt:
          type: string
          maxLength: 480
          description: Query-relevant verbatim substring of content; never model-generated.
        content:
          type: string
    DocumentSearchOutput:
      type: object
      additionalProperties: false
      required:
        - query
        - evidenceSufficient
        - results
      properties:
        query:
          type: string
        evidenceSufficient:
          type: boolean
          description: False means the Assistant must refuse or ask for clarification.
        results:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSearchResult'
...
    ErrorEnvelope:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
            - requestId
            - details
          properties:
            code:
              type: string
            message:
              type: string
            requestId:
              type: string
            details:
              type: array
              items:
                type: object
                required:
                  - reason
                properties:
                  field:
                    type: string
                  reason:
                    type: string
...
    SavedSettlementQuery:
      type: object
      additionalProperties: false
      required:
        - artistId
        - settlementMonth
      properties:
        artistId:
          type: string
          format: uuid
        settlementMonth:
          type: string
          pattern: ^[2-9]\d{3}-(?:0[1-9]|1[0-2])$
...
    MonthlyRefundMetricsRequest:
      type: object
      additionalProperties: false
      required:
        - artistId
        - settlementMonth
      properties:
        artistId:
          type: string
          format: uuid
        settlementMonth:
          type: string
          pattern: ^[2-9]\d{3}-(?:0[1-9]|1[0-2])$
    MonthlyRefundMetricsResponse:
      type: object
      additionalProperties: false
      required:
        - artistId
        - settlementMonth
        - dataCutoff
        - currency
        - refundQuantity
        - refundTransactionCount
        - refundRecordCount
        - refundAmountCents
      properties:
        artistId:
          type: string
          format: uuid
        settlementMonth:
          type: string
          pattern: ^[2-9]\d{3}-(?:0[1-9]|1[0-2])$
        dataCutoff:
          type: string
          format: date-time
        currency:
          type: string
          enum:
            - TWD
        refundQuantity:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        refundTransactionCount:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        refundRecordCount:
          type: integer
          minimum: 0
          maximum: 9007199254740991
        refundAmountCents:
          type: integer
          minimum: 0
          maximum: 9007199254740991
    ApprovalCommand:
      type: object
      additionalProperties: false
      required:
        - action
        - requestId
      properties:
        action:
          type: string
          enum:
            - submit
            - approve
            - reject
        requestId:
          type: string
          format: uuid
        reason:
          type: string
          maxLength: 2000
          default: ''
...
    ApprovalListResponse:
      type: object
      additionalProperties: false
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            type: object
            additionalProperties: false
            required:
              - reportId
              - artistId
              - displayName
              - settlementMonth
              - asOf
              - generatedAt
              - isProvisional
              - reportStatus
              - generationStatus
              - pdfFileReference
              - approvalStatus
              - trigger
              - version
              - approvalUpdatedAt
            properties:
              reportId:
                type: string
                minLength: 1
                maxLength: 120
              artistId:
                type: string
                minLength: 1
                maxLength: 120
              displayName:
                type: string
                minLength: 1
              settlementMonth:
                type: string
                pattern: ^[2-9]\d{3}-(?:0[1-9]|1[0-2])$
              asOf:
                type: string
                format: date-time
              generatedAt:
                type: string
                format: date-time
              isProvisional:
                type: boolean
              reportStatus:
                type: string
                enum:
                  - draft
              generationStatus:
                type: string
                enum:
                  - ready
              pdfFileReference:
                type: string
                minLength: 1
              approvalStatus:
                type: string
                enum: *a34
              trigger:
                type: string
                enum: *a25
              version:
                type: integer
                minimum: 0
                exclusiveMinimum: true
              approvalUpdatedAt:
                type: string
                format: date-time
                nullable: true
        nextCursor:
          type: string
          nullable: true
...